修复 torch_scatter int32 index 错误
_build_local_sub_edge_index_list 中 subgraph() 继承
graph_edge_index 的 int32 dtype，输出的窗口子图
传给 torch_scatter 时报 Expected dtype int64 for index。
窗口子图很小，输出时 .long() 转回 int64 即可。

+++ NeutronGT/core/node_level_pipeline/window_state.py
@@ -130,7 +130,8 @@ def _build_local_sub_edge_index_list(structInfo: StructInfo, local_partitions):
                 relabel_nodes=True,
                 num_nodes=structInfo.num_nodes,
             )
-            local_sub_edge_index_list.append(local_edge_index)
+            # torch_scatter 要求 int64 index; graph_edge_index 为 int32, subgraph 继承该 dtype
+            local_sub_edge_index_list.append(local_edge_index.long())
         return local_sub_edge_index_list
 
     rowptr = structInfo.graph_csr_data['rowptr'].to(torch.long).cpu()
